requirements.txt
# Core dependencies
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
cached-property==1.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
openai==1.39.0
packaging==24.0
Pillow==9.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==0.21.1
python-multipart==0.0.8
PyYAML==6.0.1
replicate==0.7.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions>=4.10.0
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets>=10.0.0
zipp==3.15.0

# New dependencies for enhanced backend (Python 3.7 compatible)
lancedb==0.5.4
motor==3.1.1
pymongo==4.3.3
web3>=6.0.0
numpy>=1.26.0
pandas>=2.1.0
networkx>=3.0
scikit-learn>=1.3.0
tiktoken>=0.5.0
langchain>=0.1.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
matplotlib>=3.7.0
pyarrow>=10.0.1

# Additional dependencies for enhanced features
reportlab>=4.0.0
sendgrid>=6.10.0
twilio>=8.0.0